refactor: create environment dir
